APIs/twilio.com/twilio_wireless_v1/1.55.0/python-aiohttp-client/tests/test_wireless_v1_rate_plan_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.list_rate_plan_response import ListRatePlanResponse
from openapi_server.models.wireless_v1_rate_plan import WirelessV1RatePlan


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("application/x-www-form-urlencoded not supported by Connexion")
async def test_create_rate_plan(client):
    """Test case for create_rate_plan

    
    """
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/x-www-form-urlencoded',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    data = {
        'data_enabled': True,
        'data_limit': 56,
        'data_metering': 'data_metering_example',
        'friendly_name': 'friendly_name_example',
        'international_roaming': ['international_roaming_example'],
        'international_roaming_data_limit': 56,
        'messaging_enabled': True,
        'national_roaming_data_limit': 56,
        'national_roaming_enabled': True,
        'unique_name': 'unique_name_example',
        'voice_enabled': True
        }
    response = await client.request(
        method='POST',
        path='/v1/RatePlans',
        headers=headers,
        data=data,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_delete_rate_plan(client):
    """Test case for delete_rate_plan

    
    """
    headers = { 
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='DELETE',
        path='/v1/RatePlans/{sid}'.format(sid='sid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_fetch_rate_plan(client):
    """Test case for fetch_rate_plan

    
    """
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/v1/RatePlans/{sid}'.format(sid='sid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_rate_plan(client):
    """Test case for list_rate_plan

    
    """
    params = [('PageSize', 56),
                    ('Page', 56),
                    ('PageToken', 'page_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/v1/RatePlans',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("application/x-www-form-urlencoded not supported by Connexion")
async def test_update_rate_plan(client):
    """Test case for update_rate_plan

    
    """
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/x-www-form-urlencoded',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    data = {
        'friendly_name': 'friendly_name_example',
        'unique_name': 'unique_name_example'
        }
    response = await client.request(
        method='POST',
        path='/v1/RatePlans/{sid}'.format(sid='sid_example'),
        headers=headers,
        data=data,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

